+++ vector3D.c
@@ -3,10 +3,10 @@
 Vector3D* Vector3D_0() {
 	return Vector3D_new_4(0, 0, 0, 1);
 }
-Vector3D* Vector3D_new_3(float x, float y, float z) {
+Vector3D* Vector3D_new_3(double x, double y, double z) {
 	return Vector3D_new_4(x, y, z, 1);
 }
-Vector3D* Vector3D_new_4(float x, float y, float z, float w) {
+Vector3D* Vector3D_new_4(double x, double y, double z, double w) {
 	Vector3D* addr = malloc(sizeof(Vector3D));
 	(*addr)[V_X] = x;
 	(*addr)[V_Y] = y;
@@ -14,9 +14,9 @@ Vector3D* Vector3D_new_4(float x, float y, float z, float w) {
 	(*addr)[V_W] = w;
 	return addr;
 }
-float dot_vectors(Vector3D *v1, Vector3D *v2) {
+double dot_vectors(Vector3D *v1, Vector3D *v2) {
 	return (*v1)[V_X] * (*v2)[V_X] + (*v1)[V_Y] * (*v2)[V_Y] + (*v1)[V_Z] * (*v2)[V_Z];
 }
 void print_vector3D(Vector3D *v) {
-	printf("[%.2f, %.2f, %.2f]\n", (*v)[V_X], (*v)[V_Y], (*v)[V_Z]);
+	printf("[%.2lf, %.2lf, %.2lf]\n", (*v)[V_X], (*v)[V_Y], (*v)[V_Z]);
 }
